server/requirements.txt
autopep8==2.0.2
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-api-logger==1.1.12
drf-yasg==1.21.5
filters-django==1.0.5
gunicorn==20.1.0
psycopg2-binary==2.9.6
pyshorteners==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
whitenoise==6.6.0


